So confused. See Description.
I can navigate to a folder from our starting directory (Desktop) but
it's not okay if I try to navigate to a folder in that new directory. I
don't know why. But if I were to try to navigate to a new folder that
lives in the old directory from the new directory...that's okay